Separate functions and classes into its own modules
