+++ flask_interface_gcp_pubsub/__init__.py
@@ -1,1 +1,2 @@
-from .views import ViewSchemeAsync, ViewSchemeBatch
+from .view_scheme_async import ViewSchemeAsync
+from .view_scheme_batch import ViewSchemeBatch
